Reference

Contents

Reference

Contents

Index

Index

JSOSolverSkeleton.SimpleUncSolver

JSOSolverSkeleton.simple_unc_solver

JSOSolverSkeleton.SimpleUncSolver — Type

simple_unc_solver(nlp; options...)

Template for an unconstrained solver using the first order conditions

‖∇f(xₖ)‖ ≤ ϵₐ + ϵᵣ‖∇f(x₀)‖,

where ϵₐ is an absolute tolerance and ϵᵣ is a relative tolerance.

For advanced usage, create a SimpleUncSolver to preallocate the memory used in the algorithm, and then call solve!.

solver = SimpleUncSolver(nlp)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

Output

The returned value is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has found a stopping criteria. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

Examples

using JSOSolverSkeleton, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
stats = simple_unc_solver(nlp)

# output

"Execution stats: first-order stationary"

using JSOSolverSkeleton, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
solver = SimpleUncSolver(nlp)
stats = solve!(solver, nlp)

# output

"Execution stats: first-order stationary"

JSOSolverSkeleton.simple_unc_solver — Method

simple_unc_solver(nlp; options...)

Template for an unconstrained solver using the first order conditions

‖∇f(xₖ)‖ ≤ ϵₐ + ϵᵣ‖∇f(x₀)‖,

where ϵₐ is an absolute tolerance and ϵᵣ is a relative tolerance.

For advanced usage, create a SimpleUncSolver to preallocate the memory used in the algorithm, and then call solve!.

solver = SimpleUncSolver(nlp)
solve!(solver, nlp; kwargs...)

Arguments

nlp::AbstractNLPModel{T, V} represents the model to solve, see NLPModels.jl.

The keyword arguments may include

x::V = nlp.meta.x0: the initial guess.

atol::T = √eps(T): absolute tolerance.

rtol::T = √eps(T): relative tolerance, the algorithm stops when ‖∇f(xᵏ)‖ ≤ atol + rtol * ‖∇f(x⁰)‖.

max_eval::Int = -1: maximum number of objective function evaluations.

max_time::Float64 = 30.0: maximum time limit in seconds.

max_iter::Int = typemax(Int): maximum number of iterations.

verbose::Int = 0: if > 0, display iteration details every verbose iteration.

Output

The returned value is a GenericExecutionStats, see SolverCore.jl.

Callback

The callback is called at each iteration. The expected signature of the callback is callback(nlp, solver, stats), and its output is ignored. Changing any of the input arguments will affect the subsequent iterations. In particular, setting stats.status = :user will stop the algorithm. All relevant information should be available in nlp and solver. Notably, you can access, and modify, the following:

solver.x: current iterate;

solver.gx: current gradient;

stats: structure holding the output of the algorithm (GenericExecutionStats), which contains, among other things:

stats.dual_feas: norm of current gradient;

stats.iter: current iteration counter;

stats.objective: current objective function value;

stats.status: current status of the algorithm. Should be :unknown unless the algorithm has found a stopping criteria. Changing this to anything will stop the algorithm, but you should use :user to properly indicate the intention.

stats.elapsed_time: elapsed time in seconds.

Examples

using JSOSolverSkeleton, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
stats = simple_unc_solver(nlp)

# output

"Execution stats: first-order stationary"

using JSOSolverSkeleton, ADNLPModels
nlp = ADNLPModel(x -> sum(x.^2), ones(3))
solver = SimpleUncSolver(nlp)
stats = solve!(solver, nlp)

# output

"Execution stats: first-order stationary"